fix : profile update

+++ src/main/java/issac/issac_server/user/application/profile/ProfileUpdater.java
@@ -7,6 +7,8 @@ import issac.issac_server.user.domain.Profile;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
+import static issac.issac_server.common.config.Constant.DEFAULT_PHOTO;
+
 @Component
 @RequiredArgsConstructor
 public class ProfileUpdater {
@@ -14,7 +16,9 @@ public class ProfileUpdater {
     private final S3Remover s3Remover;
 
     public ProfileResponse update(Profile profile, ProfileUpdateRequest request) {
-        s3Remover.deleteObjectByUrl(profile.getProfilePhotoUrl());
+        if (request.getProfilePhotoUrl() != null && !profile.getProfilePhotoUrl().endsWith(DEFAULT_PHOTO)) {
+            s3Remover.deleteObjectByUrl(profile.getProfilePhotoUrl());
+        }
         profile.update(request);
         return ProfileResponse.from(profile);
     }
